tests: Add SKIP_TEST_* variables to allow these tests to be skipped.
By setting these variables, we can skip tests that fail on
RHEL 5.
(cherry picked from commit 145f35badfd9ca4a6d9f54f7732566a5e5114876)

+++ tests/md/test-mdadm.sh
@@ -20,6 +20,11 @@
 
 set -e
 
+if [ -n "$SKIP_TEST_MDADM_SH" ]; then
+    echo "$0: test skipped because environment variable is set."
+    exit 0
+fi
+
 rm -f md-test1.img md-test2.img md-test3.img md-test4.img
 
 ../../fish/guestfish <<EOF
